Move event methods to respective module
